refactor: return empty string instead of wrapping with paragraph
