+++ packages/components/inputs/rich-text-utils/src/html/index.ts
@@ -1,1 +1,7 @@
-export { default, defaultSlateState, type Deserialized } from './html';
+export {
+  default,
+  defaultSlateState,
+  type Deserialized,
+  type Format,
+  type CustomElement,
+} from './html';
